backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
boto3==1.34.84
requests==2.31.0
groq==0.4.2
python-dotenv==1.0.1
